+++ src/utils/index.ts
@@ -9,5 +9,3 @@ export * from "./response";
 export * from "./session";
 export * from "./cors";
 export * from "./sanitize";
-
-export { ValidateFunction, ValidateResult } from "./validate";
